final solution
A game that contains a maze and a Robot that knows nothing about the maze, the exit gate, and its starting locations. Our task is developing an algorithm so that the Robot can go through the maze and reach the exit gate in the minimum steps and counting the steps that the Robot takes to go to the final destination.